fix: revert modules loading order, move ParserConfig type to common/types, add exception to injector when Parser config is not parsed in the background
