+++ src/common/types.ts
@@ -225,3 +225,38 @@ export type ContentDetector = {
   contextId: string
   selector: string
 }
+
+export type ContextBuilder = {
+  [key: string]: ContextBuilder | string
+}
+
+export type ParserConfig = {
+  themes?: {
+    DARK?: string
+    LIGHT?: string
+  }
+  contexts: {
+    [contextName: string]: {
+      containerSelector: string
+      contextSelector?: string
+      widgets?: {
+        [widgetName: string]: {
+          styles?: string
+          insertionPoint: string
+          insert?: string
+        }
+      }
+      events?: {
+        [eventName: string]: {
+          element: string
+          listen: string
+          data?: {
+            [key: string]: string
+          }
+        }
+      }
+      contextBuilder: ContextBuilder
+      childrenContexts?: string[]
+    }
+  }
+}
