--- src/contentscript/modules/config-adapter/types.ts
@@ -1,34 +1,0 @@
-export type ContextBuilder = {
-  [key: string]: ContextBuilder | string
-}
-
-export type ParserConfig = {
-  themes?: {
-    DARK?: string
-    LIGHT?: string
-  }
-  contexts: {
-    [contextName: string]: {
-      containerSelector: string
-      contextSelector?: string
-      widgets?: {
-        [widgetName: string]: {
-          styles?: string
-          insertionPoint: string
-          insert?: string
-        }
-      }
-      events?: {
-        [eventName: string]: {
-          element: string
-          listen: string
-          data?: {
-            [key: string]: string
-          }
-        }
-      }
-      contextBuilder: ContextBuilder
-      childrenContexts?: string[]
-    }
-  }
-}
